third_party/blink/web_tests/fast/forms/reportValidity-handler-updates-dom-expected.txt

HTMLFormElement::reportValidity() with cases that event handlers called by reportValidity() updates DOM structure.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

The target form is removed.
PASS $("f1").reportValidity() is false


A control to be checked is removed.
PASS $("f1").reportValidity() is false
PASS handler2Called is false
PASS document.activeElement is $("i1")


A control that was checked was removed.
PASS $("f1").reportValidity() is false
PASS handler2Called is true
PASS document.activeElement is $("i2")


A new control is added.
PASS $("f1").reportValidity() is false
PASS handler2Called is false
PASS document.activeElement is $("i1")


A control is moved to another form.
PASS $("f1").reportValidity() is false
PASS handler2Called is false
PASS document.activeElement is $("i1")


A control is moved to another document.
PASS $("f1").reportValidity() is true
PASS successfullyParsed is true

TEST COMPLETE
